+++ tools/owl-vision/CHANGELOG.md
@@ -4,7 +4,22 @@ All notable changes to the "owl-vision" extension will be documented in this fil
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
-## [Unreleased] [0.0.1] - 2023-03-10
+## [0.0.2] - 2023-2-11
+
+### Added
+
+- Switch Below command
+- Basic syntax highlight for xpaths
+- Syntax builder scripts to make syntaxes easier to read and edit
+- Syntax highlight in single quote attributes
+- Syntax highlight for slot props
+
+### Fixed
+
+- Added missing space in component's snippet indentation
+- Using `Switch Besides` or `Switch Below` does not open a new panel if one was already open
+
+## [0.0.1] - 2023-03-10
 
 - Initial release
 
@@ -15,4 +30,3 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `Find Component` Command - Finds the selected component definition.
 - `Switch` Command - Finds the corresponding template or component file depending on the current file.
 - `Switch Besides` Command - Finds the corresponding template or component file depending on the current file and opens it besides.
-- `Owl Documentation` Sidebar - A webview which allows easly searching through the owl's documentation from github